+++ project/Dependencies.scala
@@ -25,10 +25,12 @@ object Dependencies {
   val ZioLoggingVersion           = "2.5.3"
   val ZioNioVersion               = "2.0.2"
   val ZioMetricsConnectorsVersion = "2.5.5"
-  val ZioPreludeVersion           = "1.0.0-RC45"
-  val ZioSchemaVersion            = "1.7.6"
-  val ZioMockVersion              = "1.0.0-RC12"
-  val ZioVersion                  = "2.1.24"
+
+  val ZioPreludeVersion = "1.0.0-RC46"
+  val ZioSchemaVersion  = "1.8.0"
+
+  val ZioMockVersion = "1.0.0-RC12"
+  val ZioVersion     = "2.1.24"
 
   // ZIO
   val zio               = "dev.zio" %% "zio"                 % ZioVersion
@@ -106,7 +108,7 @@ object Dependencies {
   // authentication
   val jwtZioJson         = "com.github.jwt-scala" %% "jwt-zio-json" % "11.0.3"
   val springSecurityCore =
-    "org.springframework.security" % "spring-security-core" % "7.0.2" exclude (
+    "org.springframework.security" % "spring-security-core" % "7.0.3" exclude (
       "commons-logging",
       "commons-logging",
     ) exclude ("org.springframework", "spring-aop")
@@ -126,7 +128,7 @@ object Dependencies {
   val scalaCsv       = "com.github.tototoshi" %% "scala-csv"        % "2.0.0"
 
   // test
-  val dataFaker = "net.datafaker" % "datafaker" % "2.5.3"
+  val dataFaker = "net.datafaker" % "datafaker" % "2.5.4"
 
   val scalaTest = "org.scalatest" %% "scalatest" % "3.2.19"
 
@@ -137,7 +139,7 @@ object Dependencies {
   // found/added by the plugin but deleted anyway
   val commonsLang3 = "org.apache.commons" % "commons-lang3" % "3.20.0"
 
-  val tapirVersion = "1.13.6"
+  val tapirVersion = "1.13.7"
 
   val tapir = Seq(
     "com.softwaremill.sttp.tapir" %% "tapir-zio-http-server"   % tapirVersion,
@@ -151,7 +153,7 @@ object Dependencies {
     "com.softwaremill.sttp.tapir" %% "tapir-zio-metrics"                 % tapirVersion,
   )
 
-  val otelVersion = "1.58.0"
+  val otelVersion = "1.59.0"
 
   val openTelemetryWithSentry = Seq(
     "com.softwaremill.sttp.client4" %% "opentelemetry-tracing-zio-backend"   % SttpClientVersion,
@@ -205,12 +207,12 @@ object Dependencies {
     zio,
   ) ++ zioSttpClient ++ metrics ++ tapir ++ openTelemetryWithSentry
 
-  val flywayVersion        = "11.19.1"
+  val flywayVersion        = "11.20.3"
   val otelAgentVersion     = "v2.18.1"
   val otelPyroscopeVersion = "v1.0.4"
   val hikariVersion        = "7.0.2"
   val quillVersion         = "4.8.6"
-  val sqliteVersion        = "3.51.1.0"
+  val sqliteVersion        = "3.51.1.1"
 
   val db = Seq(
     "org.xerial"   % "sqlite-jdbc"    % sqliteVersion,