+++ webapi/src/main/scala/org/knora/webapi/slice/api/v3/resources/ResourcesRestServiceV3.scala
@@ -7,7 +7,6 @@ package org.knora.webapi.slice.api.v3.resources
 import zio.*
 
 import org.knora.webapi.messages.StringFormatter
-import org.knora.webapi.slice.admin.domain.model.KnoraProject
 import org.knora.webapi.slice.admin.domain.model.KnoraProject.ProjectIri
 import org.knora.webapi.slice.admin.domain.service.KnoraProjectService
 import org.knora.webapi.slice.api.v3.NotFound
@@ -28,21 +27,33 @@ class ResourcesRestServiceV3(
 
   def resourcesPerOntology(projectIri: ProjectIri): IO[V3ErrorInfo, List[OntologyAndResourceClasses]] =
     for {
-      prj        <- projectService.findById(projectIri).orDie.someOrFail(NotFound(projectIri))
-      ontologies <- ontologyRepo.findByProject(projectIri).orDie
-      result     <- ZIO
-                  .foreach(ontologies.map(o => o.ontologyIri -> o.resourceClassIris).toList)((ontoIri, classes) =>
-                    for {
-                      onto <- ontologyRestService.asOntologyDto(ontoIri)
-                      rcls <- ZIO.foreachPar(classes)(asResourceClassAndCount(prj)).withParallelism(5)
-                    } yield OntologyAndResourceClasses(onto, rcls),
-                  )
+      prj         <- projectService.findById(projectIri).orDie.someOrFail(NotFound(projectIri))
+      ontologies  <- ontologyRepo.findByProject(projectIri).orDie
+      allClasses   = ontologies.flatMap(_.resourceClassIris).toList
+      classGroups <- ZIO.foreach(allClasses)(resolveClassGroup).map(_.toMap).orDie
+      allIris      = classGroups.values.flatten.toList.distinct
+      countsByIri <- resourcesRepo.countByResourceClasses(allIris, prj).orDie
+      result      <- ZIO.foreach(ontologies.map(o => o.ontologyIri -> o.resourceClassIris).toList)((ontoIri, classes) =>
+                  for {
+                    onto <- ontologyRestService.asOntologyDto(ontoIri)
+                    rcls <- ZIO.foreach(classes)(iri =>
+                              for {
+                                rc   <- ontologyRestService.asResourceClassDto(iri)
+                                group = classGroups.getOrElse(iri, List(iri))
+                                count = group.flatMap(countsByIri.get).sum
+                              } yield ResourceClassAndCountDto(rc, count),
+                            )
+                  } yield OntologyAndResourceClasses(onto, rcls),
+                )
     } yield result
 
-  private def asResourceClassAndCount(prj: KnoraProject)(iri: ResourceClassIri) = for {
-    resourceClass <- ontologyRestService.asResourceClassDto(iri)
-    count         <- resourcesRepo.countByResourceClass(iri, prj).orDie
-  } yield ResourceClassAndCountDto(resourceClass, count)
+  private def resolveClassGroup(iri: ResourceClassIri) = for {
+    subclasses  <- ontologyRepo.findAllSubclassesBy(iri)
+    ontoIri      = iri.ontologyIri.smartIri.toInternalSchema
+    sameOntoSubs = subclasses
+                     .filter(_.entityInfoContent.classIri.getOntologyFromEntity.toInternalSchema == ontoIri)
+                     .map(sc => ResourceClassIri.unsafeFrom(sc.entityInfoContent.classIri))
+  } yield iri -> (iri :: sameOntoSubs)
 }
 
 object ResourcesRestServiceV3 {